Infer sample purity using a mixture of Binomial or Beta-Binomial distributions.

Infer sample purity using a mixture of Binomial or Beta-Binomial distributions.

estimate_purity(x, model = "Binomial", eps = 0.01)

Arguments

x
An object of class 'TAPACLOTH' generated with function init.

model
The expected distribution for the number of reads with variant at fixed coverage and purity, that can be chosen between Binomial (no over-dispersion) or Beta-Binomial (over-dispersion included).

eps
In case data is fitted with a mixture of 2 distributions, sample purity is estimated using only that whose peak is closer to the input purity estimate than the value of this parameter. If both distributions satisfy this condition, the one which is closest to the input purity estimate is used for the new estimate.

Value

An object of class TAPACLOTH containing the inferred purity.

Examples

input = init(mutations = example_data$data, sample = example_data$sample, purity = example_data$purity)
out = estimate_purity(x = input, model = "binomial", eps = 0.01)
#> 
#> ── TAPACLOTH purity estimate of sample P-0050275-T01-IM6 using binomial model ──
#> 
#> 
#> ── BMix fit ────────────────────────────────────────────────────────────────────
#> 
#> ℹ Binomials k_B = 1, 2, and 3, Beta-Binomials k_BB = 0; 6 fits to run.
#> 
#> ℹ Bmix best fit completed in 0 mins
#> 
#> ── [ BMix ] My BMix model n = 10 with k = 3 component(s) (3 + 0) ───────────────
#> • Clusters: π = 40% [Bin 2], 40% [Bin 1], and 20% [Bin 3], with π > 0.
#> • Binomial Bin 1 with mean = 0.214805305896057.
#> • Binomial Bin 2 with mean = 0.453596300416989.
#> • Binomial Bin 3 with mean = 0.56500898736096.
#> ℹ Score (model selection): ICL = 153.42.
#> Warning: `guides(<scale> = FALSE)` is deprecated. Please use `guides(<scale> = "none")` instead.
#> Warning: Removed 6 rows containing missing values (geom_bar).
#> Warning: Graphs cannot be horizontally aligned unless the axis parameter is set. Placing graphs unaligned.
print(out)
#> ── [ TAPACLOTH ]  Input data for sample P-0050275-T01-IM6, with purity 0.9 ─────
#> # A tibble: 10 × 10
#>    chr        from        to ref   alt   gene      NV    DP   VAF gene_role     
#>    <chr>     <int>     <int> <chr> <chr> <chr>  <int> <int> <dbl> <chr>         
#>  1 chr17   7578542   7578542 C     G     TP53     397   850 0.467 oncogene, TSG…
#>  2 chr12  25380275  25380275 A     C     KRAS     634  1121 0.566 oncogene      
#>  3 chr9   21971120  21971120 C     T     CDKN2A   389   866 0.449 TSG           
#>  4 chr9   21971120  21971120 C     T     CDKN2A   389   866 0.449 TSG           
#>  5 chr1   27105553  27105553 C     T     ARID1A   309   548 0.564 TSG, fusion   
#>  6 chr9   21971120  21971120 C     T     CDKN2A   389   866 0.449 TSG           
#>  7 chr1   46724370  46724370 C     T     RAD54L    76   520 0.146 NA            
#>  8 chr17  30315354  30315354 G     A     SUZ12     93   542 0.172 oncogene, TSG…
#>  9 chr2  158622618 158622618 A     G     ACVR1    175   616 0.284 oncogene      
#> 10 chr6     394875    394875 C     T     IRF4     158   659 0.240 oncogene, TSG…
#> ── [ TAPACLOTH ]  Purity estimate using  [ BMix ]  with binomial model ─────────
#> # A tibble: 1 × 3
#>   purity reliability purity_bmix
#>    <dbl>       <dbl>       <dbl>
#> 1    0.9       0.989        0.91
#>