Classify mutations using a (Beta-)Binomial model-based test.

Classify mutations using a (Beta-)Binomial model-based test.

Usage

run_classifier(
  x,
  cutoff = 0.75,
  model = "Binomial",
  rho = 0.01,
  karyotypes = c("1:0", "1:1", "2:0", "2:1", "2:2")
)

Arguments

x

An object of class 'TAPACLOTH' generated with function init.

cutoff

Likelihood cut-off for class assignment.

model

Model used for the classification task, either "Binomial" (no over-dispersion) or "Beta-Binomial" (over-dispersion included), that will be used as the expected distribution for the number of reads with variant at fixed coverage and purity.

rho

If "Beta-Binomial" model is selected, this parameter tunes the over-dispersion of the expected distribution used for the test.

karyotypes

Karyotypes to be included among the possible classes.

Value

An object of class TAPACLOTH containing the input plus the classification data and paramters.

Examples

x = init(mutations = example_data$data,
         sample = example_data$sample,
         purity = example_data$purity)
x = run_classifier(
    x = x, 
    model = "Beta-Binomial", 
    rho = 0.01,
    karyotypes = c("1:0","1:1","2:0","2:1","2:2")
    )
#> 
#> ── TAPACLOTH beta-binomial clonality/Zygosity testing for sample P-0050275-T01-I
#> 
#> ℹ Computing null model distributions and p-values.
#> Warning: numerical expression has 3 elements: only the first used
#> Warning: numerical expression has 3 elements: only the first used
#> Warning: numerical expression has 3 elements: only the first used
#> Warning: numerical expression has 3 elements: only the first used
#> Warning: numerical expression has 3 elements: only the first used
#> Warning: numerical expression has 3 elements: only the first used
#> Warning: numerical expression has 3 elements: only the first used
#> Warning: numerical expression has 3 elements: only the first used
#> Warning: numerical expression has 3 elements: only the first used
#> Warning: longer object length is not a multiple of shorter object length
#> Warning: numerical expression has 3 elements: only the first used
#> Warning: numerical expression has 3 elements: only the first used
#> Warning: numerical expression has 3 elements: only the first used
#> Warning: numerical expression has 3 elements: only the first used
#> Warning: numerical expression has 3 elements: only the first used
#> Warning: numerical expression has 3 elements: only the first used
#> Warning: numerical expression has 3 elements: only the first used
#> Warning: numerical expression has 3 elements: only the first used
#> Warning: numerical expression has 3 elements: only the first used
#> Warning: longer object length is not a multiple of shorter object length
#> Warning: numerical expression has 3 elements: only the first used
#> Warning: numerical expression has 3 elements: only the first used
#> Warning: numerical expression has 3 elements: only the first used
#> Warning: numerical expression has 3 elements: only the first used
#> Warning: numerical expression has 3 elements: only the first used
#> Warning: numerical expression has 3 elements: only the first used
#> Warning: numerical expression has 3 elements: only the first used
#> Warning: numerical expression has 3 elements: only the first used
#> Warning: numerical expression has 3 elements: only the first used
#> Warning: longer object length is not a multiple of shorter object length
print(x)
#> ── [ TAPACLOTH ]  Input data for sample P-0050275-T01-IM6, with purity 0.9 ─────
#> # A tibble: 10 × 10
#>    chr        from        to ref   alt   gene      NV    DP   VAF gene_role     
#>    <chr>     <int>     <int> <chr> <chr> <chr>  <int> <int> <dbl> <chr>         
#>  1 chr17   7578542   7578542 C     G     TP53     397   850 0.467 oncogene, TSG…
#>  2 chr12  25380275  25380275 A     C     KRAS     634  1121 0.566 oncogene      
#>  3 chr9   21971120  21971120 C     T     CDKN2A   389   866 0.449 TSG           
#>  4 chr9   21971120  21971120 C     T     CDKN2A   389   866 0.449 TSG           
#>  5 chr1   27105553  27105553 C     T     ARID1A   309   548 0.564 TSG, fusion   
#>  6 chr9   21971120  21971120 C     T     CDKN2A   389   866 0.449 TSG           
#>  7 chr1   46724370  46724370 C     T     RAD54L    76   520 0.146 NA            
#>  8 chr17  30315354  30315354 G     A     SUZ12     93   542 0.172 oncogene, TSG…
#>  9 chr2  158622618 158622618 A     G     ACVR1    175   616 0.284 oncogene      
#> 10 chr6     394875    394875 C     T     IRF4     158   659 0.240 oncogene, TSG…
#> ── [ TAPACLOTH ]  Test using beta-binomial model with overdispersion parameter 0
#> # A tibble: 10 × 9
#>    id                        NV    DP   VAF gene  gene_…¹ ploidy multi…² uncer…³
#>    <chr>                  <int> <int> <dbl> <chr> <chr>    <int>   <int>   <dbl>
#>  1 chr17:7578542:7578542…   397   850 0.467 TP53  oncoge…      4       2  0.494 
#>  2 chr12:25380275:253802…   634  1121 0.566 KRAS  oncoge…      3       2  0.352 
#>  3 chr9:21971120:2197112…   389   866 0.449 CDKN… TSG          2       1  0.482 
#>  4 chr9:21971120:2197112…   389   866 0.449 CDKN… TSG          2       1  0.482 
#>  5 chr1:27105553:2710555…   309   548 0.564 ARID… TSG, f…      3       2  0.396 
#>  6 chr9:21971120:2197112…   389   866 0.449 CDKN… TSG          2       1  0.482 
#>  7 chr1:46724370:4672437…    76   520 0.146 RAD5… NA           4       1  0.0106
#>  8 chr17:30315354:303153…    93   542 0.172 SUZ12 oncoge…      4       1  0.0295
#>  9 chr2:158622618:158622…   175   616 0.284 ACVR1 oncoge…      3       1  0.399 
#> 10 chr6:394875:394875:C:T   158   659 0.240 IRF4  oncoge…      4       1  0.264 
#> # … with abbreviated variable names ¹​gene_role, ²​multiplicity, ³​uncertainty